+++ rubato/utils/__init__.py
@@ -1,20 +1,18 @@
 """
 This module houses low-dependency utilities widely used in rubato.
 """
-from .path import *
+from .path import get_path
 from .error import *
+from .rb_time import DelayedTask, FramesTask, RecurrentTask, Time
 from .rb_math import Math
 from .noise import Noise
 from .vector import Vector
 from .display import Display
 from .color import Color
-from .surface import Surface
-from .font import Font
-from .rb_input import Input
+from .rendering import *
 from .events import *
 from .radio import Radio, Listener
 from .camera import Camera
 from .draw import Draw, DrawTask
-from .rb_time import DelayedTask, FramesTask, ScheduledTask, Time
 from .sound import Sound
 from .debug import Debug
