+++ rubato/utils/hardware/sound.py
@@ -9,7 +9,7 @@ from warnings import warn
 import sdl2.sdlmixer as mixer
 from sdl2 import AUDIO_F32
 
-from . import IdError, get_path, Math
+from .. import IdError, get_path, Math
 
 if mixer.Mix_OpenAudio(48000, AUDIO_F32, 2, 2048):
     warn("Could not open audio device.")
