+++ rubato/utils/rendering/camera.py
@@ -5,7 +5,7 @@ GameObjects only render if their z-index is not more than that of the camera's.
 
 The current scene's camera can be accessed through :code:`Game.camera`.
 """
-from . import Vector, Display, Math
+from .. import Vector, Display, Math
 
 
 class Camera:
