+++ rubato/utils/rendering/font.py
@@ -4,7 +4,7 @@ import sdl2, sdl2.sdlttf, sdl2.ext
 from importlib.resources import files
 import ctypes
 
-from . import Color
+from .. import Color
 
 
 class Font:
