+++ rubato/utils/rendering/surface.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from typing import Optional
 import sdl2, sdl2.ext, ctypes
 
-from ..c_src import c_draw
-from . import Vector, Color, Display, get_path
+from ...c_src import c_draw
+from .. import Vector, Color, Display, get_path
 
 
 class Surface:
